+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 QuantumCircuitOpt.jl Change Log
 ===============================
 
+### v0.5.4
+- Added a generalized function to obtain controlled gates (`controlled_gate()`) in any number of qubits
+- Cleaned-up controlled gates in `src/gates.jl`
+- Updated docs to reflect above changes
+
 ### v0.5.3
 - Minor update: SC22 publication added in docs
 - README banner update